A lot of general fixes (see description)
1) Fixed soundtrack volume;
2) Adapted all std::cout messages with "DEBUG:" prefix;
3) Introduced a lot of forceGameClosure and showGameWarning (file existing or integrity check);
4) Fixed some warning conversion errors (a lot more left by the way);
5) Added error codes into a json file;
6) Added text translations;

+++ centurion-source/source/audio/audio_player.h
@@ -42,7 +42,7 @@ public:
 	virtual bool IsLooping();
 
 	virtual void SetPitch(float value);
-	virtual void SetVolume(float value);
+	virtual void SetVolume(int value);
 
 	ALuint GetSourceID()
 	{
